11 спринт 2 ревью

+++ src/utils/api.js
@@ -1,7 +1,7 @@
-import {apiCofig} from './utils';
+import { apiCofig } from "./utils";
 
-class Api{
-  constructor({url,headers}){
+class Api {
+  constructor({ url, headers }) {
     this._baseUrl = url;
     this._token = headers;
   }
@@ -15,53 +15,49 @@ class Api{
       );
     }
   }
-// Получения информации о пользователе
-  getUserInfo(){
-  return fetch(`${this._baseUrl}users/me`, {
-  headers: this._token
-})
-  .then((res) =>{
-    return this._requestResult(res);
-  })
-}
-//Получение карточек 
- getInitialCard(){
-  return fetch (`${this._baseUrl}cards`,{
-  headers: this._token
-})
-.then((res) =>{
-  return this._requestResult(res);
-  })
-}
-//Запрос на изменение профиля
-  editProfile({name,about}){
-    return fetch(`${this._baseUrl}users/me`,{
+  // Получения информации о пользователе
+  getUserInfo() {
+    return fetch(`${this._baseUrl}users/me`, {
+      headers: this._token,
+    }).then((res) => {
+      return this._requestResult(res);
+    });
+  }
+  //Получение карточек
+  getInitialCard() {
+    return fetch(`${this._baseUrl}cards`, {
+      headers: this._token,
+    }).then((res) => {
+      return this._requestResult(res);
+    });
+  }
+  //Запрос на изменение профиля
+  editProfile(name, about) {
+    return fetch(`${this._baseUrl}users/me`, {
       method: "PATCH",
       headers: this._token,
       body: JSON.stringify({
         name: name,
         about: about,
-    }),
-    })
-    .then(res =>{
+      }),
+    }).then((res) => {
       return this._requestResult(res);
-  })
-}
-// Запрос на изменение аватара
-  editAvatar({avatar}){
-    return fetch(`${this._baseUrl}users/me/avatar`,{
+    });
+  }
+  // Запрос на изменение аватара
+  editAvatar(avatar) {
+    return fetch(`${this._baseUrl}users/me/avatar`, {
       method: "PATCH",
       headers: this._token,
       body: JSON.stringify({
-        avatar: avatar, 
+        avatar: avatar,
       }),
-    })
-    .then(res =>{
+    }).then((res) => {
       return this._requestResult(res);
-    })
+    });
   }
 
-  addCard({ name, link }) {
+  addCard(name, link) {
     return fetch(`${this._baseUrl}cards`, {
       method: "POST",
       headers: this._token,
@@ -71,35 +67,32 @@ class Api{
       }),
     }).then(this._requestResult);
   }
-  
-// Поставновка лайка
-  setLikes(idCard){
-    return fetch(`${this._baseUrl}cards/${idCard}/likes`,{
-      method: "PUT",
-      headers: this._token,
-    })
-    .then(res =>{
-      return this._requestResult(res);
-    })
+
+  // Постановка и удалание лайка
+  changeLikeCardStatus(idCard, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._baseUrl}cards/${idCard}/likes`, {
+        method: "PUT",
+        headers: this._token,
+      }).then((res) => {
+        return this._requestResult(res);
+      });
+    } else {
+      return fetch(`${this._baseUrl}cards/${idCard}/likes`, {
+        method: "DELETE",
+        headers: this._token,
+      }).then((res) => {
+        return this._requestResult(res);
+      });
+    }
   }
-// Удаление лайка
-  deleteLikes(idCard){
-    return fetch(`${this._baseUrl}cards/${idCard}/likes`,{
-      method: "DELETE",
-      headers: this._token,
-    })
-    .then(res =>{
-      return this._requestResult(res);
-    })
-  }
-  
+
   deleteCard(id) {
     return fetch(`${this._baseUrl}cards/${id}`, {
       method: "DELETE",
       headers: this._token,
-    })
-    .then((res)=>{
-      this._requestResult(res)
+    }).then((res) => {
+      this._requestResult(res);
     });
   }
 }
